:fire: refactor(controllers): reorganize student, course, and progress logic
This substantial refactor reorganizes student-facing course and progress
logic into new, dedicated controllers, centralizing responsibilities and
slimming down legacy controllers.

- **Introduce `PublicCourseController`**: Takes over `index` (course
catalog) and `show` (course details) logic from the now-deleted
`CourseController`, handling public, unauthenticated, and initial course
viewing.
- **Introduce `StudentCourseController`**: Takes over `learn` (course
consumption), `complete` (manual course completion), and `viewPurchased`
(purchased courses list) logic from the deleted `CourseController` and
`PurchasedCoursesController`. Also absorbs `updateProgress` from the
deleted `CourseProgressController`. This centralizes all authenticated
student course interaction.
- **Delete Controllers**:
    - `CourseController` (Functionality moved to
`PublicCourseController` and `StudentCourseController`).
    - `PurchasedCoursesController` (Functionality moved to
`StudentCourseController`).
    - `CourseProgressController` (Functionality moved to
`StudentCourseController`).
- **Controller Refinements**:
    - `ChatController`: Remove the unused `index` method, leaving only
`show` and messaging methods, as chat contact logic is likely moved
elsewhere or handled client-side.
    - `CourseCartController`: Rename `show` to `index` to align with
REST conventions.
    - `TutorController`: Remove unused standard REST methods (`index`,
`create`, `store`, `edit`, `update`, `destroy`), leaving only `show`.
    - `ForumController`: Minor style fixes (consistent use of braces and
semicolons in `switch` statements).
    - `ProfileController`: Cleanup of commented-out, unused methods and
use of `auth()` helper.
    - `Tutor/CourseLessonController` and
`Tutor/CourseSectionController`: Use `$user->id()` consistently for
authorization checks instead of direct `auth()->id()`.
- **Routing Updates**: Update `routes/web.php` and `CourseDetails.jsx`
to point to the new, reorganized controller methods and paths. Remove
the defunct `chat.index` route.
