+++ app/Http/Controllers/ProfileController.php
@@ -3,25 +3,25 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\ProfileUpdateRequest;
-use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Http\RedirectResponse;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Redirect;
 use Inertia\Inertia;
-use Inertia\Response;
 
 class ProfileController extends Controller
 {
-    public function edit() {
+    public function edit()
+    {
+        $user = auth();
+        /** @var \App\Models\User $user */
+
         return Inertia::render('Profile', [
             'auth' => [
-                'user' => auth()->user()->load('role'),
+                'user' => $user->user()->load('role'),
             ],
         ]);
     }
 
-    public function update(ProfileUpdateRequest $request) : RedirectResponse {
+    public function update(ProfileUpdateRequest $request): RedirectResponse
+    {
         $user = $request->user();
 
         if ($request->hasFile('profileImage')) {
@@ -38,52 +38,4 @@ class ProfileController extends Controller
 
         return back()->with('success', 'Profile updated!');
     }
-
-    /**
-     * Display the user's profile form.
-     */
-    // public function edit(Request $request): Response
-    // {
-    //     return Inertia::render('Profile/Edit', [
-    //         'mustVerifyEmail' => $request->user() instanceof MustVerifyEmail,
-    //         'status' => session('status'),
-    //     ]);
-    // }
-
-    /**
-     * Update the user's profile information.
-     */
-    // public function update(ProfileUpdateRequest $request): RedirectResponse
-    // {
-    //     $request->user()->fill($request->validated());
-
-    //     if ($request->user()->isDirty('email')) {
-    //         $request->user()->email_verified_at = null;
-    //     }
-
-    //     $request->user()->save();
-
-    //     return Redirect::route('profile.edit');
-    // }
-
-    /**
-     * Delete the user's account.
-     */
-    // public function destroy(Request $request): RedirectResponse
-    // {
-    //     $request->validate([
-    //         'password' => ['required', 'current_password'],
-    //     ]);
-
-    //     $user = $request->user();
-
-    //     Auth::logout();
-
-    //     $user->delete();
-
-    //     $request->session()->invalidate();
-    //     $request->session()->regenerateToken();
-
-    //     return Redirect::to('/');
-    // }
 }
